secondary piece in pianoroll, move gpt2 logic to a module
