+++ gpt2/configs/model/gpt2.yaml
