+++ gpt2/configs/model/gpt2_minimal.yaml
