+++ gpt2/configs/model/gpt2_scaled.yaml
